requirements.txt
torch>=2.0.0
transformers>=4.36.0
pillow>=9.0.0
pillow-heif>=1.0.0
accelerate>=0.24.0
protobuf>=3.20.0
tqdm>=4.64.0
click>=8.0.0
colorama>=0.4.6
